LayoutTests/fast/selectors/querySelector-explicit-case-insensitive-attribute-match-with-svg-expected.txt

Some attribute needs case insensitive value matching with HTML, they should not match XML elements like SVG. On the other hand, explicitely case-insensitive selector should match SVG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("[lang=foobar]").length is 3
PASS document.querySelectorAll("[lang=foobar]")[0].id is "div1"
PASS document.querySelectorAll("[lang=foobar]")[1].id is "div3"
PASS document.querySelectorAll("[lang=foobar]")[2].id is "rect3"
PASS document.querySelectorAll("[lang=foobar i]").length is 4
PASS document.querySelectorAll("[lang=foobar i]")[0].id is "div1"
PASS document.querySelectorAll("[lang=foobaR i]")[1].id is "div3"
PASS document.querySelectorAll("[lang=Foobar i]")[2].id is "rect1"
PASS document.querySelectorAll("[lang=FOOBAR i]")[3].id is "rect3"
PASS document.querySelectorAll("[link=foobar]").length is 3
PASS document.querySelectorAll("[link=foobar]")[0].id is "div2"
PASS document.querySelectorAll("[link=foobar]")[1].id is "div4"
PASS document.querySelectorAll("[link=foobar]")[2].id is "rect4"
PASS document.querySelectorAll("[link=foobar i]").length is 4
PASS document.querySelectorAll("[link=foobar i]")[0].id is "div2"
PASS document.querySelectorAll("[link=FOOBAR i]")[1].id is "div4"
PASS document.querySelectorAll("[link=foobaR i]")[2].id is "rect2"
PASS document.querySelectorAll("[link=Foobar i]")[3].id is "rect4"
PASS successfullyParsed is true

TEST COMPLETE
